Cleanup BackendService after generation, add omitted fields.
